build: Update to koka v3.2.2
Switch from using the `--optimize=2` option to `--ccopts="-O2"`. The
newer version of koka will try to pass `-mtune` and `-march` options to
emcc which will cause the wasm build to fail.

Add the `-sSTANDALONE_WASM=1` option to better communicate that the
generated wasm is intended to run without javascript support.